requirements.txt
absl-py==2.1.0
altair==5.2.0
amqp==5.2.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.11
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==4.2.0
black==20.8b1
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.3
celery==5.3.6
certifi==2020.12.5
cffi==1.17.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confusable-homoglyphs==3.2.0
contourpy==1.2.0
cryptography==43.0.0
cssselect2==0.7.0
cycler==0.10.0
Cython==0.29.21
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.5
dlib==19.21.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.8.0
flatbuffers==24.3.25
fonttools==4.49.0
future==1.0.0
gitdb==4.0.11
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
identify==2.5.6
idna==2.10
importlib_metadata==8.2.0
importlib_resources==6.1.3
imutils==0.5.3
iniconfig==2.0.0
inside==0.0.4
ipykernel==6.29.5
ipynb==0.5.1
ipython==8.18.1
isort==5.10.1
jax==0.4.26
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kombu==5.3.5
lazy-object-proxy==1.7.1
lxml==5.3.0
maestrowf==1.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.1.1
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.7.0
notebook==7.2.0
notebook-as-pdf==0.5.0
notebook_shim==0.2.4
numpy==2.0.1
oauthlib==3.2.1
opencv-contrib-python==4.9.0.80
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.1
pandas==2.2.1
pandoc==2.3
pandocfilters==1.5.1
parse==1.20.1
parso==0.8.4
pathspec==0.8.1
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.4.0
plumbum==1.8.3
ply==3.11
pre-commit==2.20.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.5
pure_eval==0.2.3
pvrecorder==1.2.2
pyarrow==15.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.8.1b0
pydyf==0.9.0
pyee==11.1.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.5.0
pylint==2.15.4
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==3.0.1
pyperclip==1.9.0
pyphen==0.14.0
pyppeteer==2.0.0
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenrec==0.4
PyScreeze==0.1.30
pytest==8.0.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytweening==1.2.0
pytz==2022.6
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.1.0
razorpay==1.4.2
redis==5.0.3
referencing==0.33.0
regex==2020.11.13
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SoundCard==0.4.3
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.28
sqlparse==0.4.3
stack-data==0.6.3
streamlit==1.32.0
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.3.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typed-ast==1.4.2
types-cryptography==3.3.23
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.4
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.20.0
zopfli==0.2.3
